requirements.txt
PySide6>=6.0.0
pyfingerprint
pyserial
tensorflow
Pillow
numpy
